Basic working AMQP spout.
Give it a broker address, the name of an exchange to bind to, and a
routing key to select messages, and it'll emit a tuple for each message
published to the exchange.

Currently the tuples it emits just contain one element, which is the
string obtained by decoding the bytes of the message in the default
encoding (probably UTF-8).